[qtdocgallery] Provide directory mimetype for file rootType models. JB#55184 OMP#JOLLA-287
Tracker:FileSystem graph doesn't anymore provide mimeTypes except
indirectly for directories. (Path maps to urn with nie:interpretedAs, and
the urn has the mimeType)

With such little usage, originally the mimeType was skipped for this type,
but granted that it's quite useful to know what items are files and what
are directories, so added some special handling.